feat(configure): add VS Code MCP client support (#125)
Implement VS Code client integration for MCP configuration.

I also added `upateConfig` and `hasExistingConfig` methods to the base
client interface. This allows us to defer the detection (i.e. if the
server is already configured) and the actual updating logic (i.e.
writing to the config file) to the underlying `MCPClientConfig`
implementation. This is largely because VSCode's structure is pretty
different from the others. In order to avoid annoying duplication, I've
left the default implementation in the base client (shared by cursor,
windsurf, claude), and then override that in the VSCode client.