feat(local-mcp-server): add custom user-agent header
The `SDK_METADATA` for `@gleanwork/api-client` looks like:

```json
{
  language: 'typescript',
  openapiDocVersion: '0.9.0',
  sdkVersion: '0.6.1',
  genVersion: '2.610.0',
  userAgent: 'speakeasy-sdk/typescript 0.6.1 2.610.0 0.9.0 @gleanwork/api-client'
}
```

This PR introduces a custom `User-Agent` string specifically for
`@gleanwork/local-mcp-server`. It is very similar to the default, but it
customizes the version (second field) to be the published package
version and the package name (last field) to be
`@gleanwork/local-mcp-server`.